requirements.txt
appdirs==1.4.4
asn1crypto==1.4.0
astroid==2.4.2
Authlib==0.15.3
autopep8==1.5.5
awswrangler==2.7.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
boto3==1.17.53
botocore==1.20.53
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
contextvars==2.4
cryptography==3.4.4
cssselect==1.1.0
cycler==0.10.0
Django==1.11.29
et-xmlfile==1.0.1
fake-useragent==0.1.11
feedparser==6.0.2
flake8==3.8.4
Flask==1.1.2
Flask-Cors==3.0.10
h11==0.12.0
html5lib==1.1
httpcore==0.12.3
httpx==0.16.1
idna==2.10
immutables==0.15
importlib-metadata==2.1.1
install==1.3.4
isort==5.7.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.3
jmespath==0.10.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
lxml==4.6.2
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
numpy==1.18.1
openpyxl==3.0.7
pandas==1.1.5
pandas-cache==0.0.1
paramiko==2.7.2
parse==1.19.0
parso==0.7.1
pg8000==1.18.0
Pillow==8.1.0
pluggy==0.13.1
psutil==5.8.0
psycopg2==2.8.6
pyarrow==3.0.0
pycodestyle==2.6.0
pycparser==2.20
pydocstyle==5.1.1
pyee==8.1.0
pyflakes==2.2.0
pylint==2.6.0
PyMySQL==1.0.2
PyNaCl==1.4.0
pyparsing==2.4.7
pyppeteer==0.2.5
pyquery==1.4.3
pyreadline==2.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
pytz==2021.1
redshift-connector==2.0.878
requests==2.25.1
requests-html==0.10.0
rfc3986==1.4.0
rope==0.18.0
s3transfer==0.3.7
scipy==1.5.4
scramp==1.2.2
selenium==3.141.0
sgmllib3k==1.0.0
six==1.14.0
sniffio==1.2.0
snowballstemmer==2.1.0
soupsieve==2.2
SQLAlchemy==1.3.23
tda-api==1.1.5
timedelta==2020.12.3
toml==0.10.2
tqdm==4.56.2
typed-ast==1.4.2
ujson==4.0.2
urllib3==1.26.3
w3lib==1.22.0
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
yahoo-earnings-calendar==0.6.0
yahoo-fin==0.8.8
yapf==0.30.0
zipp==3.4.0
